+++ Changes.md
@@ -41,6 +41,7 @@ Documentation
 - Updated the "Macbeth Chart" example to use Cycles instead of Appleseed.
 - Updated the "Per-location Light Tweak Spreadsheet" example to use Cycles instead of Appleseed.
 - Updated the "Wedge Tests" example to use Cycles instead of Appleseed.
+- Updated the "Setting Up a Spreadsheet" tutorial to use Cycles instead of Appleseed.
 
 1.2.2.0 (relative to 1.2.1.1)
 =======